refactor auth/user code into existing structure
