+++ mondey_backend/src/mondey_backend/databases/users.py
@@ -4,32 +4,20 @@ from collections.abc import AsyncGenerator
 from typing import Annotated
 
 from fastapi import Depends
-from fastapi_users.db import SQLAlchemyBaseUserTable
 from fastapi_users.db import SQLAlchemyUserDatabase
-from sqlalchemy import Integer
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.ext.asyncio import async_sessionmaker
 from sqlalchemy.ext.asyncio import create_async_engine
-from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy.orm import Mapped
-from sqlalchemy.orm import mapped_column
 
-from . import MONDEY_DATABASE_PATH
-
-
-class Base(DeclarativeBase):
-    pass
-
-
-class User(SQLAlchemyBaseUserTable[int], Base):
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-
+from .. import MONDEY_DATABASE_PATH
+from ..models.users import Base
+from ..models.users import User
 
 engine = create_async_engine(f"sqlite+aiosqlite:///{MONDEY_DATABASE_PATH}/users.db")
 async_session_maker = async_sessionmaker(engine, expire_on_commit=False)
 
 
-async def create_db_and_tables():
+async def create_user_db_and_tables():
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
