+++ mondey_backend/src/mondey_backend/models/users.py
@@ -1,0 +1,28 @@
+from __future__ import annotations
+
+from fastapi_users import schemas
+from fastapi_users.db import SQLAlchemyBaseUserTable
+from sqlalchemy import Integer
+from sqlalchemy.orm import DeclarativeBase
+from sqlalchemy.orm import Mapped
+from sqlalchemy.orm import mapped_column
+
+
+class Base(DeclarativeBase):
+    pass
+
+
+class User(SQLAlchemyBaseUserTable[int], Base):
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+
+
+class UserRead(schemas.BaseUser[int]):
+    pass
+
+
+class UserCreate(schemas.BaseUserCreate):
+    pass
+
+
+class UserUpdate(schemas.BaseUserUpdate):
+    pass
