+++ mondey_backend/src/mondey_backend/routers/users.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from fastapi import APIRouter
+
+from ..models.users import UserRead
+from ..models.users import UserUpdate
+from ..users import fastapi_users
+
+router = APIRouter(prefix="/users", tags=["users"])
+router.include_router(fastapi_users.get_users_router(UserRead, UserUpdate))
