--- mondey_backend/src/mondey_backend/schemas.py
@@ -1,15 +1,0 @@
-from __future__ import annotations
-
-from fastapi_users import schemas
-
-
-class UserRead(schemas.BaseUser[int]):
-    pass
-
-
-class UserCreate(schemas.BaseUserCreate):
-    pass
-
-
-class UserUpdate(schemas.BaseUserUpdate):
-    pass
